requirements.txt
Flask==1.1.2
nltk==3.6.2
pandas==1.2.4
python-dotenv==0.17.1
requests=2.25.1
regex==2021.4.4
scikit-learn==0.24.2
gunicorn==20.1.0